migrate build code for phoenix cart
