--- conf.d/downloads
@@ -1,8 +1,0 @@
-#!/bin/bash -ex
-
-dl() {
-    [ "$FAB_HTTP_PROXY" ] && PROXY="--proxy $FAB_HTTP_PROXY"
-    cd $2; curl -w "%{filename_effective}" -L -f -OJs $PROXY $1; cd -
-}
-
-dl http://www.oscommerce.com/download-file /usr/local/src
