+++ overlay/etc/apache2/sites-available/phoenix_cart.conf
@@ -11,14 +11,14 @@ ServerName localhost
 <VirtualHost *:443>
     SSLEngine on
     ServerAdmin webmaster@localhost
-    DocumentRoot /var/www/oscommerce/
+    DocumentRoot /var/www/phoenix_cart/
     RewriteEngine On
     # redirect to configured domain or IP (updated by inithook)
     RewriteCond %{HTTP_HOST} !^localhost$
     RewriteRule ^/(.*) https://localhost/$1 [R=307,L]
 </VirtualHost>
 
-<Directory /var/www/oscommerce/>
+<Directory /var/www/phoenix_cart/>
     Options +FollowSymLinks
     AllowOverride All
     Require all granted
