--- overlay/usr/lib/inithooks/firstboot.d/20regen-oscommerce-secrets
@@ -1,15 +1,0 @@
-#!/bin/bash -e
-# regenerate oscommerce mysql password
-
-. /etc/default/inithooks
-
-PASSWORD=$(mcookie)
-
-CONF=/var/www/oscommerce/includes/local/configure.php
-sed -i "s|define('DB_SERVER_PASSWORD',.*|define('DB_SERVER_PASSWORD', '$PASSWORD');|" $CONF
-
-CONF=/var/www/oscommerce/admin/includes/local/configure.php
-sed -i "s|define('DB_SERVER_PASSWORD',.*|define('DB_SERVER_PASSWORD', '$PASSWORD');|" $CONF
-
-$INITHOOKS_PATH/bin/mysqlconf.py --user=oscommerce --pass="$PASSWORD"
-
