+++ overlay/usr/lib/inithooks/firstboot.d/20regen-phoenix-cart-secrets
@@ -1,0 +1,15 @@
+#!/bin/bash -e
+# regenerate phoenix cart mysql password
+
+. /etc/default/inithooks
+
+PASSWORD=$(mcookie)
+
+CONF=/var/www/phoenix_cart/includes/configure.php
+sed -i "s|DB_SERVER_PASSWORD = .*|DB_SERVER_PASSWORD = '$PASSWORD';|" $CONF
+
+CONF=/var/www/phoenix_cart/administration/includes/configure.php
+sed -i "s|DB_SERVER_PASSWORD = .*|DB_SERVER_PASSWORD = '$PASSWORD';|" $CONF
+
+$INITHOOKS_PATH/bin/mysqlconf.py --user=phoenix_cart --pass="$PASSWORD"
+
