--- overlay/usr/lib/inithooks/firstboot.d/40oscommerce
@@ -1,8 +1,0 @@
-#!/bin/bash -e
-# set oscommerce admin password and email
-
-. /etc/default/inithooks
-
-[ -e $INITHOOKS_CONF ] && . $INITHOOKS_CONF
-$INITHOOKS_PATH/bin/oscommerce.py --pass="$APP_PASS" --email="$APP_EMAIL" --domain="$APP_DOMAIN"
-
